moved BookingRepository
from 'interfaces' to 'data'
